+++ 017io.cc
@@ -97,5 +97,5 @@ COMPILE_FN(read-byte, compiledFn_read_byte, "('$eof)",
     return mkref(lookup("$eof"));
   char c;
   f.read(&c, 1);
-  return mkref(newNum((int)c));
+  return mkref(newNum((long)c));
 )
